challenge/backend/requirements.txt
anyio==3.7.0
ariadne==0.19.1
blinker==1.6.2
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pyee==9.0.4
PyJWT==2.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.2
Werkzeug==2.3.4
